# loms/pkg/api/order/order.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Order"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/order/cancel": {
      "post": {
        "operationId": "Order_Cancel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderOrderCancelRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/order/create": {
      "post": {
        "operationId": "Order_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderOrderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderOrderCreateRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/order/info/{order_id}": {
      "get": {
        "operationId": "Order_Info",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderOrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/order/pay": {
      "post": {
        "operationId": "Order_Pay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderOrderPayRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    }
  },
  "definitions": {
    "orderOrderCancelRequest": {
      "type": "object",
      "properties": {
        "order_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "orderOrderCreateRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/orderOrderCreateRequestItem"
          }
        }
      }
    },
    "orderOrderCreateRequestItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "orderOrderCreateResponse": {
      "type": "object",
      "properties": {
        "order_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "orderOrderInfoResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/orderOrderInfoResponseItem"
          }
        }
      }
    },
    "orderOrderInfoResponseItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "orderOrderPayRequest": {
      "type": "object",
      "properties": {
        "order_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "x-auth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  }
}